1.a: ah
